% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Return the mode of numeric vector}
\usage{
mode(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric
}
\description{
Return the mode of numeric vector
}
\examples{

\dontrun{

library(cousepkg)

x <- c(1,2,3,4,4,5,5,5)

mode(x)

}

}
