% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_couse.R
\name{theme_couse}
\alias{theme_couse}
\title{Set pretty defaults for ggplot2 objects}
\usage{
theme_couse(
  base_size = 12,
  strip_text_size = 12,
  strip_text_margin = 5,
  subtitle_size = 13,
  subtitle_margin = 10,
  plot_title_size = 16,
  plot_title_margin = 10,
  plot_caption_size = 10,
  axis_x_margin = 10,
  axis_y_margin = 5,
  axis_text_size = 12,
  legend_title_size = 12,
  legend_text_size = 10,
  fill = "steelblue3",
  ...
)
}
\arguments{
\item{base_size}{base font size}

\item{strip_text_size}{strip text size}

\item{strip_text_margin}{bottom margin for strip title}

\item{subtitle_size}{subtitle size}

\item{subtitle_margin}{bottom margin for subtitle}

\item{plot_title_size}{plot title size}

\item{plot_title_margin}{bottom margin for plot title}

\item{axis_x_margin}{top margin for x axis title}

\item{axis_y_margin}{right margin for y axis title}

\item{fill}{default fill color for geom_bar, geom_rect, and geom_col}

\item{...}{other args to \code{theme_minimal}}
}
\value{
A ggplot2 theme
}
\description{
Set pretty defaults for ggplot2 objects
}
\examples{

\dontrun{

library(tidyverse)

theme_set(theme_couse())
}


}
